requirements.txt
firebase-admin==6.2.0
functions-framework==3.3.0
google-cloud-storage==2.9.0
pymavlink==2.4.37
arweave-python-client==1.0.3
solana==0.30.2
anchorpy==0.14.0
cryptography==41.0.3
requests==2.31.0
aiohttp==3.8.5
async-timeout==4.0.3
pycryptodome==3.19.0
